Restructures and refactors API endpoints
Refactors the API structure by renaming and moving files for better organization.

Updates the API configuration to a dedicated config.json file and modifies the application name.

Adjusts API routes to provide more descriptive information and interact with external services like Flickr.
