+++ app/api/gl-api/route.tsx
@@ -1,13 +1,16 @@
-import config from './response.json';
+// core/app/api/gl-api/route.ts
+import config from './config.json';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(request: NextRequest) {
-  const { app = 'App Title', endpoints, frontend } = config;
+  const { app = 'Goldlabel' } = config;
 
   return NextResponse.json({
     time: Date.now(),
-    frontend,
     app,
-    endpoints,
+    endpoint: "gl-api",
+    title: `${app} API`,
+    flickr: "http://localhost:3000/api/gl-api/flickr"
   });
 }
+
